Fix issues based on comments
