--- module/book_module.rb
@@ -1,24 +1,0 @@
-module BookDisplay
-  def list_books
-    @books.each do |book|
-      puts "Title: '#{book.title}', Author: #{book.author}"
-    end
-  end
-
-  def list_people
-    @people.each do |person|
-      puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-
-    print 'Author: '
-    author = gets.chomp
-
-    @books << Book.new(title, author)
-    puts 'Book created successfully'
-  end
-end
