--- module/rental_module.rb
@@ -1,46 +1,0 @@
-module RentalDisplay
-  def create_rental
-    create_rental_book
-    book_index = gets.chomp.to_i
-    puts
-    create_rental_person
-    person_index = gets.chomp.to_i
-    puts
-    print 'Date: '
-    date = gets.chomp
-    @rentals << Rental.new(date, @books[book_index], @people[person_index])
-    puts 'Rental created successfully'
-  end
-
-  def create_rental_book
-    if @books.empty?
-      puts 'There are no books in the library to rent'
-      return
-    end
-    puts 'Select a book from the following list by number'
-    @books.each_with_index do |book, index|
-      puts "#{index}) Title: '#{book.title}', Author: #{book.author}"
-    end
-  end
-
-  def create_rental_person
-    puts 'Select a person from the following list by number (not id)'
-    if @people.empty?
-      puts 'There are no people in the library'
-      return
-    end
-
-    @people.each_with_index do |person, index|
-      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-  end
-
-  def list_rentals_for_person_id
-    print 'ID of person: '
-    id = gets.chomp.to_i
-    puts 'Rentals:'
-    @rentals.each do |rental|
-      puts "Date: #{rental.date}, Book '#{rental.book.title}' by #{rental.book.author}" if rental.person.id == id
-    end
-  end
-end
